+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
  * Enable mocking non-shared services in tests
  * Add `KernelTestCase::executeCommand()` to execute a command in functional tests
  * Add `framework.html_sanitizer.sanitizers.*.default_action` config option
+ * Deprecate parameters `router.request_context.scheme` and `router.request_context.host`;
+   use the `router.request_context.base_url` parameter or the `framework.router.default_uri` config option instead
 
 8.0
 ---